........
  r22270 | wes.young | 2012-12-11 19:15:05 -0700 (Tue, 11 Dec 2012) | 1 line
  
  Daily VERSION for SVN tree
........
  r22271 | sandra.castro | 2012-12-12 02:02:26 -0700 (Wed, 12 Dec 2012) | 12 lines
  
  
         New Development: Yes
              JIRA Issue: CAS-4755
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: 
           Test Programs: test_mstransform
    Put in Release Notes: No
               Module(s): mstransform
             Description: First implementation of task mstransform and tool mstransformer, which is able to combine spws for one input MS only. The test is failing in the same way that the tMSTransformDataHanler test is failing. This will be fixed shortly.
  	
........
  r22272 | sandra.castro | 2012-12-12 08:08:18 -0700 (Wed, 12 Dec 2012) | 12 lines
  
  
         New Development: No
              JIRA Issue: CAS-4765
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: 
           Test Programs: test_mstransform
    Put in Release Notes: No
               Module(s): mstransform
             Description: Added the datacolumn parameter in MSTransformDataHandler::configure() and changed the default of columname_p to 'CORRECTED'. Made datacolumn was uppercase inside MSTransform::configure().
  	
........
  r22274 | kumar.golap | 2012-12-12 11:32:48 -0700 (Wed, 12 Dec 2012) | 15 lines
  
  
         New Development: No
              JIRA Issue: Yes CAS-4764
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: Please list interface changes
           Test Programs: List test programs
    Put in Release Notes: Yes/No
               Module(s): Module Names change impacts.
             Description: Describe your changes here...
  
  Fixed a bug that  happens when the stop button is pressed in interactive window
  	
........
  r22277 | jeff.kern | 2012-12-12 14:32:06 -0700 (Wed, 12 Dec 2012) | 14 lines
  
  
         New Development: No
              JIRA Issue: No
          Ready for Test: No
       Interface Changes: No
  What Interface Changed: Please list interface changes
           Test Programs: List test programs
    Put in Release Notes: No
               Module(s): Synthesis
             Description: Cleanup to remove all "trivial" warnings.
  
  	
........
  r22283 | michel.caillat | 2012-12-12 16:51:49 -0700 (Wed, 12 Dec 2012) | 14 lines
  
  
         New Development: No
              JIRA Issue: No
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: 
           Test Programs: 
    Put in Release Notes: No
               Module(s): alma ASDMBinaries
             Description: A first effort to reduce the number of compilation warnings in this directory. 
  
  	
........
  r22284 | george.moellenbrock | 2012-12-12 17:18:27 -0700 (Wed, 12 Dec 2012) | 16 lines
  
  
         New Development: No
              JIRA Issue: No
          Ready for Test: Yes
       Interface Changes: Yes
  What Interface Changed: removed various deprecated/unneeded interfaces
           Test Programs: regressions...
    Put in Release Notes: No
               Module(s): synthesis
             Description: Code cleanup related to the NewCalTable migration,
  			as well as some excision of old, unneeded 
  			methods in various calibration-related classes.
  			This is in preparation for migration to the new
  			VI/VB framework.
........
  r22289 | michel.caillat | 2012-12-13 08:41:19 -0700 (Thu, 13 Dec 2012) | 14 lines
  
  
         New Development: No
              JIRA Issue: No
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: 
           Test Programs: 
    Put in Release Notes: 
               Module(s): alma ASDMBinaries
             Description: A second effort to reduce the number of compilation warnings. 1) Completed the copy ctors of some derived class by calling their super classes ctors 2) Replaced stat64 and fstat64 by stat and fstat to avoid the deprecated warning (hope this wont create any problem, since we are all in 64bits)
  
  	
........
  r22290 | susan.loveland | 2012-12-13 10:19:57 -0700 (Thu, 13 Dec 2012) | 15 lines
  
  
         New Development: No
              JIRA Issue: No
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: Please list interface changes
           Test Programs: List test programs
    Put in Release Notes: Yes/No
               Module(s): Module Names change impacts.
             Description: Work on histogram fuctionality.  Fixed bug in
  Feather where a file separator was not being added to a directory.
  
  	
........
  r22291 | kumar.golap | 2012-12-13 11:01:27 -0700 (Thu, 13 Dec 2012) | 15 lines
  
  
         New Development: No
              JIRA Issue: No List JIRA ticket.
          Ready for Test: Yes
       Interface Changes: Yes/No
  What Interface Changed: Please list interface changes
           Test Programs: List test programs
    Put in Release Notes: Yes/No
               Module(s): Module Names change impacts.
             Description: Describe your changes here...
  
  Moved history saving to restored image to iClean as restoring of images has moved there.
  	
........
  r22292 | susan.loveland | 2012-12-13 11:16:13 -0700 (Thu, 13 Dec 2012) | 15 lines
  
  
         New Development: No
              JIRA Issue: No
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: Please list interface changes
           Test Programs: List test programs
    Put in Release Notes: No
               Module(s): Module Names change impacts.
             Description: Update of example histogram code in viewer, and making
  hitogram plotMode enum public.
  
  	
........
  r22294 | dave.mehringer | 2012-12-13 12:20:14 -0700 (Thu, 13 Dec 2012) | 14 lines
  
  
         New Development: No
              JIRA Issue: CAS-4767 
          Ready for Test: Yes
       Interface Changes: Yes/No
  What Interface Changed: Please list interface changes
           Test Programs: List test programs
    Put in Release Notes: Yes/No
               Module(s): Module Names change impacts.
             Description: moved ImageAnalysis from casacore
  
  	
........
  r22296 | tak.tsutsumi | 2012-12-13 15:12:18 -0700 (Thu, 13 Dec 2012) | 14 lines
  
  
         New Development: No
              JIRA Issue: Yes CAS-4606
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: Please list interface changes
           Test Programs: test_setjy 
    Put in Release Notes: Yes/No
               Module(s): Module Names change impacts.
             Description: Added/removed some comments.
  
  	
........
  r22298 | urvashi.rao | 2012-12-13 15:59:14 -0700 (Thu, 13 Dec 2012) | 14 lines
  
  
         New Development: No
              JIRA Issue: cas-2671
          Ready for Test: Yes
       Interface Changes: No
  What Interface Changed: Please list interface changes
           Test Programs: List test programs
    Put in Release Notes: No
               Module(s): Module Names change impacts.
             Description: Indexing off-by-1 fix from Sanjay
  
  	
........


git-svn-id: https://svn.cv.nrao.edu/svn/casa/test@22334 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/synthesis/TransformMachines/accumulateFromGrid.cc
@@ -23,13 +23,13 @@
 
     for(Int iy=-scaledSupport[1]; iy <= scaledSupport[1]; iy++) 
       {
-	iLoc[1]    = (scaledSampling[1]*iy+offset[1]);
+	iLoc[1]    = static_cast<int>(scaledSampling[1]*iy+offset[1]);
 	iCFPos[1]  = iLoc[1] + convOrigin[1];
 	iGrdPos[1] = loc[1]+iy;
 
 	for(Int ix=-scaledSupport[0]; ix <= scaledSupport[0]; ix++) 
 	  {
-	    iLoc[0]    = (scaledSampling[0]*ix+offset[0]);
+	    iLoc[0]    = static_cast<int>(scaledSampling[0]*ix+offset[0]);
 	    iCFPos[0]  = iLoc[0] + convOrigin[0];
 	    iGrdPos[0] = loc[0]+ix;
